fix: set margin-bottom value for all major headers
Previous to this change, `margin-bottom` was set to zero for `h3`-level headers.

In some cases, the header overlapped with a codeblock thereby cutting off part of the text.

This change explicitly sets a `margin-bottom` value for `h1`, `h2`, and `h3` headers.

In addition, it groups common attributes together (like `font-weight` and `line-height`) and brings `font-size` in their targeted groups.